Changed: double/Double to BigDecimal

+++ src/main/java/org/bilanzius/persistence/models/Transaction.java
@@ -1,14 +1,16 @@
 package org.bilanzius.persistence.models;
 
+import java.math.BigDecimal;
+
 public class Transaction {
 
-    public static Transaction create(User user, BankAccount account, Category category, double money, String description) {
+    public static Transaction create(User user, BankAccount account, Category category, BigDecimal money, String description) {
         return new Transaction(
                 0, user.getId(), account.getAccountId(), category.getCategoryId(), money, description
         );
     }
 
-    public static Transaction create(User user, BankAccount account, double money, String description) {
+    public static Transaction create(User user, BankAccount account, BigDecimal money, String description) {
         return new Transaction(
                 0, user.getId(), account.getAccountId(), -1, money, description // -1 as default categoryId
         );
@@ -18,11 +20,11 @@ public class Transaction {
     private final int userId;
     private final int accountId;
     private final int categoryId;
-    private final double money;
+    private final BigDecimal money;
 
     private final String description;
 
-    public Transaction(int transactionId, int userId, int accountId, int categoryId, double money, String description) {
+    public Transaction(int transactionId, int userId, int accountId, int categoryId, BigDecimal money, String description) {
         this.transactionId = transactionId;
         this.userId = userId;
         this.accountId = accountId;
@@ -47,7 +49,7 @@ public class Transaction {
         return categoryId;
     }
 
-    public double getMoney() {
+    public BigDecimal getMoney() {
         return money;
     }
